requirements.txt
datasets>=2.0.0
torch>=1.10.0
transformers>=4.20.0
Pillow>=8.0.0
sentencepiece>=0.1.99
requests>=2.25.1
tqdm>=4.65.0 